+++ src/main/kotlin/cloud/glitchdev/rfu/utils/TextUtils.kt
@@ -14,6 +14,10 @@ object TextUtils {
         return Component.literal("§b§l[§f§lRFU§b§l] $textColor$string")
     }
 
+    fun rfupfLiteral(string: String, textStyle: TextStyle) : MutableComponent {
+        return Component.literal("§b§l[§f§lRFU§9§lPF§b§l] $textStyle$string")
+    }
+
     fun rfupfLiteral(string: String, textColor: TextColor = TextColor.WHITE) : MutableComponent {
         return Component.literal("§b§l[§f§lRFU§9§lPF§b§l] $textColor$string")
     }