+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Fixed Hud elements disappearing after moving them
 - Party finder now reloads upon opening
 - Fixed party finder not correctly tracking player count on new parties
+- Fixed party join msgs not being hidden
 
 ### Back-end:
 - Refactor to add an AbstractTextHudElement for simple text hud stuff